update todolist board
