+++ client/src/stylesheets/components/_dashboard.scss
@@ -22,6 +22,7 @@
   align-items: center;
 }
 
+
 //dashboard
 .dashboard {
   @include align-flex-item(column);
