+++ client/src/stylesheets/components/_projectboard.scss
@@ -1,47 +1,4 @@
-.board-body{
-	background: $body-color
+.board-body {
+  background: $body-color;
 }
 
-.project-list{
-	background: #202040;
-	padding-bottom: 5px;
-	h3{
-		color: $theme-green;
-		font-family: $roboto-mono;
-	}
-	hr{
-		height: 0.01em;
-    width: 65vw;
-    color: #ecf4f3;
-    background: #ecf4f3;
-		margin-left: unset;
-		margin-bottom: unset;
-	}
- 
-	span{
-		@include space-between;
-	}
-
-	div{
-		color: white;
-		margin-right: 10px;
-		span{
-			margin-left: 7px;
-			font-size: 10px;
-		}
-	}
-}
-
-.todo-icon{
-	border: 1px $theme-green solid;
-	font-size: 25px;
-	padding: 5px;
-	color: white;
-}
-
-.summary-cards{
-	.summary-icon{
-		flex-basis: 10%;
-		background: rgba(128, 128, 128, 0.479);
-	}
-}
