+++ client/src/stylesheets/components/_todolist.scss
@@ -1,0 +1,76 @@
+.todo-body{
+	position: relative;
+}
+
+.project-list {
+  background: #202040;
+	padding-bottom: 5px;
+	// border: red 3px solid;
+	@include drop-shadow(rgba(0, 0, 0, 0.425));
+  h3 {
+    color: $theme-green;
+    font-family: $roboto-mono;
+  }
+  hr {
+    height: 0.01em;
+    width: 65vw;
+    color: #ecf4f3;
+    background: #ecf4f3;
+    margin-left: unset;
+    margin-bottom: unset;
+  }
+
+  span {
+    @include space-between;
+  }
+
+  div {
+    color: rgba(255, 255, 255, 0.466);
+    margin-right: 10px;
+    span {
+      margin-left: 7px;
+      font-size: 10px;
+    }
+	}
+	
+	.date{
+	 color: rgba(255, 255, 255, 0.466);
+	 font-size: 10px;
+	}
+}
+
+.todo-icon {
+  border: 1px $theme-green solid;
+  font-size: 25px;
+  padding: 5px;
+  color: white;
+}
+
+.summary-cards {
+  justify-content: unset;
+  .summary-icon {
+    flex-basis: 10%;
+    background: rgba(128, 128, 128, 0.479);
+  }
+  div:nth-child(2) {
+    margin-right: 7px;
+    cursor: pointer;
+  }
+
+  .clicked {
+    background: $theme-green;
+    color: white;
+	}
+	
+}
+
+// close todo button
+.close-todo{
+	background: rgba(128, 128, 128, 0.322);
+	height: 20px;
+	color: rgba(255, 255, 255, 0.63);
+	padding: 5px;
+	position: absolute;
+	right: 0;
+	cursor: pointer;
+}
