+++ client/src/stylesheets/main.scss
@@ -1,4 +1,5 @@
 @import './components/dashboard';
+@import './components/todolist';
 @import './components/card';
 @import './components/addProject';
 @import './components/projectboard';
